+++ libnopegl/src/node_drawtexture.c
@@ -33,7 +33,6 @@
 #include "node_block.h"
 #include "node_texture.h"
 #include "pipeline.h"
-#include "transforms.h"
 #include "utils/darray.h"
 #include "utils/memory.h"
 #include "utils/utils.h"
@@ -79,7 +78,6 @@ struct pipeline_desc {
     struct ngli_pipeline *pipeline;
     NGLI_DARRAY(struct resource_map) blocks_map;
     NGLI_DARRAY(struct texture_map) textures_map;
-    struct ngli_node_darray reframing_nodes;
 };
 
 #define TEXTURE_WRAP_DEFAULT 0
@@ -148,7 +146,7 @@ static const float default_uvcoords[] = {
 #define OFFSET(x) offsetof(struct drawtexture_opts, x)
 static const struct node_param drawtexture_params[] = {
     {"texture",  NGLI_PARAM_TYPE_NODE, OFFSET(texture_node),
-                 .node_types=(const uint32_t[]){TRANSFORM_TYPES_ARGS, NGL_NODE_TEXTURE2D, NGL_NODE_CUSTOMTEXTURE, NGLI_NODE_NONE},
+                 .node_types=(const uint32_t[]){NGL_NODE_TEXTURE2D, NGL_NODE_CUSTOMTEXTURE, NGLI_NODE_NONE},
                  .flags=NGLI_PARAM_FLAG_NON_NULL,
                  .desc=NGLI_DOCSTRING("texture to render")},
     {"wrap",     NGLI_PARAM_TYPE_SELECT, OFFSET(wrap),
@@ -173,12 +171,6 @@ static int drawtexture_init(struct ngl_node *node)
     const struct drawtexture_opts *o = node->opts;
     struct ngl_ctx *ctx = node->ctx;
     struct ngpu_ctx *gpu_ctx = ctx->gpu_ctx;
-
-    const struct ngl_node *texture_node = ngli_transform_get_leaf_node(o->texture_node);
-    if (!texture_node) {
-        LOG(ERROR, "no texture found at the end of the transform chain");
-        return NGL_ERROR_INVALID_USAGE;
-    }
 
     if (!o->geometry) {
         s->own_geometry = 1;
@@ -290,7 +282,7 @@ static int drawtexture_prepare(struct ngl_node *node,
     ngli_assert(vert_size == sizeof(struct drawtexture_vert_block));
     ngli_assert(frag_size >= sizeof(struct drawtexture_frag_block));
 
-    const struct ngl_node *texture_node = ngli_transform_get_leaf_node(o->texture_node);
+    const struct ngl_node *texture_node = o->texture_node;
     struct texture_info *texture_info = texture_node->priv_data;
     struct ngpu_pgcraft_texture textures[] = {
         {
@@ -396,9 +388,6 @@ static int drawtexture_prepare(struct ngl_node *node,
             return NGL_ERROR_MEMORY;
     }
 
-    /* Push reframing node for the texture */
-    if (ngli_darray_try_push(&desc->reframing_nodes, o->texture_node) < 0)
-        return NGL_ERROR_MEMORY;
 
     return 0;
 }
@@ -454,13 +443,9 @@ static void drawtexture_draw(struct ngl_node *node)
     struct texture_map *texture_map = desc->textures_map.data;
     for (size_t i = 0; i < desc->textures_map.count; i++) {
         if (texture_map[i].image_rev != texture_map[i].image->rev) {
-            ngli_pipeline_update_image(pipeline, (int32_t)i, texture_map[i].image, ctx->current_staging_buffer);
+            ngli_pipeline_update_image(pipeline, (int32_t)i, texture_map[i].image);
             texture_map[i].image_rev = texture_map[i].image->rev;
         }
-
-        struct ngli_mat4 reframing_matrix = {0};
-        ngli_transform_chain_compute(desc->reframing_nodes.data[i], reframing_matrix.m);
-        ngli_pipeline_apply_reframing_matrix(pipeline, (int32_t)i, texture_map[i].image, reframing_matrix.m, ctx->current_staging_buffer);
     }
 
     struct resource_map *resource_map = desc->blocks_map.data;
@@ -484,9 +469,9 @@ static void drawtexture_draw(struct ngl_node *node)
     if (s->geometry->indices_buffer) {
         const struct ngpu_buffer *indices = s->geometry->indices_buffer;
         const struct buffer_layout *layout = &s->geometry->indices_layout;
-        ngli_pipeline_draw_indexed(pipeline, indices, layout->format, (uint32_t)layout->count, 1);
+        ngli_pipeline_draw_indexed(pipeline, ctx->current_staging_buffer, indices, layout->format, (uint32_t)layout->count, 1);
     } else {
-        ngli_pipeline_draw(pipeline, s->nb_vertices, 1, 0);
+        ngli_pipeline_draw(pipeline, ctx->current_staging_buffer, s->nb_vertices, 1, 0);
     }
 }
 
@@ -499,7 +484,6 @@ static void drawtexture_uninit(struct ngl_node *node)
     ngli_pipeline_freep(&desc->pipeline);
     ngli_darray_reset(&desc->blocks_map);
     ngli_darray_reset(&desc->textures_map);
-    ngli_darray_reset(&desc->reframing_nodes);
 
     /* Free crafter and block descriptors */
     ngpu_pgcraft_freep(&s->crafter);